Creation des seeders

+++ database/migrations/2024_07_15_125222_create_filieres_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->uuid('id')->primary();
             $table->string('nom_filiere')->unique();
             $table->text('description')->nullable();
-            $table->float('montant_formation');
+            $table->decimal('montant_formation', 10, 2); // Précision de 15 chiffres au total, 2 après la virgule
             $table->timestamps();
         });
     }
